script: fix error that didn't remove all tweets
There was some occurences where some tweets by people inside the bots
array were not being removed.

After some effort trying to find the cause, it was determined that this
was happening because of the forEach loop inside the tweet, where if
there was any result after the bot, it was overwritten, leaving the
final result as false.

The bug fix was done by creating a flag that marks if at any moment
a bot was found, in this case, the entire tweet is removed.

Signed-off-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>